# openapi.yaml
openapi: 3.1.0
info:
  title: PSRU AI Virtual Photo Booth & VR Studio API
  version: "1.0.0"
  description: >
    REST API for the PSRU AI Virtual Photo Booth & VR Studio.
    Capture → AI render (segmentation, scene generation, relighting, outfit) →
    branding → output. PDPA-aware consent gating and audit trail.
servers:
  - url: https://api.photobooth.psru.ac.th/v1
security:
  - bearerAuth: []
...
paths:
...
  /jobs:
    post:
      tags: [Jobs]
      summary: Create a render job
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [capture_id, scene_id]
              properties:
                capture_id: { type: string, format: uuid }
                scene_id: { type: string, format: uuid }
                outfit_id: { type: string, format: uuid }
                fx: { type: object, additionalProperties: true }
                branding_id: { type: string, format: uuid }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "422": { description: Rejected by content guardrail }
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job status
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /outputs/{id}/download:
    get:
      tags: [Outputs]
      summary: Get a signed download URL
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: fmt, in: query, schema: { type: string, enum: [jpg, png, tiff, pdf] } }
      responses:
        "302": { description: Redirect to signed CDN URL }
  /outputs/{id}/share:
    post:
      tags: [Outputs]
      summary: Create a share token / QR
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  share_token: { type: string }
                  qr_url: { type: string }
                  expires_at: { type: string, format: date-time }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Job:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, succeeded, failed, canceled] }
        progress: { type: integer }
        stage: { type: string }
        output_id: { type: string, format: uuid }
...